requirements.txt
numpy==2.1.3
pandas==2.2.3
pipenv==2024.3.1
python-dotenv==1.0.1
requests==2.32.3
